refactor: deepen runtime target seams
